[Feature][Test][API] 用户测试类新增 注入测试接口
