+++ mosaic-agent/dependency-reduced-pom.xml
@@ -51,6 +51,12 @@
   </build>
   <dependencies>
     <dependency>
+      <groupId>com.alibaba</groupId>
+      <artifactId>fastjson</artifactId>
+      <version>2.0.47</version>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
       <groupId>io.github.classgraph</groupId>
       <artifactId>classgraph</artifactId>
       <version>4.8.179</version>
@@ -90,12 +96,6 @@
       <groupId>org.javassist</groupId>
       <artifactId>javassist</artifactId>
       <version>3.29.2-GA</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.alibaba</groupId>
-      <artifactId>fastjson</artifactId>
-      <version>1.2.83</version>
       <scope>provided</scope>
     </dependency>
   </dependencies>
