+++ mosaic-core/src/main/java/io/github/tml/mosaic/core/doc/core/AbstractTmlDoc.java
@@ -1,4 +1,4 @@
-package io.github.tml.mosaic.doc.core;
+package io.github.tml.mosaic.core.doc.core;
 
 import java.util.Collection;
 import java.util.List;
