+++ mosaic-core/src/main/java/io/github/tml/mosaic/core/doc/features/extend/TmValidation.java
@@ -1,7 +1,7 @@
-package io.github.tml.mosaic.doc.features.extend;
+package io.github.tml.mosaic.core.doc.features.extend;
 
-import io.github.tml.mosaic.doc.core.TmlDoc;
-import io.github.tml.mosaic.doc.part.Validation;
+import io.github.tml.mosaic.core.doc.core.TmlDoc;
+import io.github.tml.mosaic.core.doc.part.Validation;
 import java.util.Map;
 import java.util.Optional;
