+++ mosaic-core/src/main/java/io/github/tml/mosaic/core/doc/features/extend/TmValue.java
@@ -1,7 +1,7 @@
-package io.github.tml.mosaic.doc.features.extend;
+package io.github.tml.mosaic.core.doc.features.extend;
 
-import io.github.tml.mosaic.doc.core.TmlDoc;
-import io.github.tml.mosaic.doc.part.Value;
+import io.github.tml.mosaic.core.doc.core.TmlDoc;
+import io.github.tml.mosaic.core.doc.part.Value;
 
 import java.util.Map;
 import java.util.Optional;
